requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
django-admin-rangefilter==0.9.0
django-ckeditor==6.5.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
phonenumbers==8.12.56
Pillow==9.2.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==1.5
pytz==2023.3
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
whitenoise==6.2.0
